+++ ec3730/Views/HostView.swift
@@ -5,31 +5,6 @@ import Combine
 import Introspect
 import InterposeKit
 
-struct FSDisclosureGroup<Label, Content>: View where Label: View, Content: View {
-  @Binding var isExpanded: Bool
-  var content: () -> Content
-  var label: () -> Label
-
-  var body: some View {
-    VStack(alignment: .leading, spacing: 0.0) {
-        Button(action: {
-            withAnimation {
-                self.isExpanded.toggle()
-            }
-        }, label: {
-            HStack(alignment: .center) {
-                label()
-                Image(systemName: "chevron.down").rotationEffect(self.isExpanded ? .degrees(0.0) : .degrees(-90.0)).padding()
-            }
-        })
-        
-        if self.isExpanded {
-            content()
-        }
-    }
-  }
-}
-
 @available(iOS 15.0, *)
 struct HostView: View {
     @EnvironmentObject var model: HostViewModel
